Upgrade to freeimage 3.18.0 (Fixes #18)

+++ src/freeimage/Source/OpenEXR/IlmImf/ImfOptimizedPixelReading.h
@@ -70,7 +70,7 @@ EXR_FORCEINLINE
 bool
 isPointerSSEAligned (const void* EXR_RESTRICT pPointer)
 {
-    unsigned long trailingBits = ((intptr_t)pPointer) & 15;
+    unsigned long trailingBits = ((unsigned long)pPointer) & 15;
     return trailingBits == 0;
 }
